sec: gateway receives join request with edhoc m1
